user can update status

+++ rest-api/controllers/feed.js
@@ -1,5 +1,6 @@
 const { validationResult } = require('express-validator/check'); 
 const Post = require('../models/post'); 
+const User = require('../models/user');
 const fs = require('fs');
 const path = require('path'); 
 
@@ -59,22 +60,34 @@ exports.createPost = (req, res, next) => {
   const imageUrl = req.file.path;
   const title = req.body.title;
   const content = req.body.content; 
+  let creator;
   // Create post in database 
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: { name: 'Maximilian' }
+    creator: req.userId
   }); 
   // Save post in database
   post.save()
     .then(result => {
-      console.log(result); 
+      // Find the matching user
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      creator = user;
+      // add the currently created post to the user   
+      user.posts.push(post);
+      return user.save()
+    })
+    .then(result => {
+      console.log(result);
       console.log('CREATED POST SUCCESSFULLY')
       // Send response to client 
       res.status(201).json({
         message: 'Post created successfully',
-        post: result
+        post: post,
+        creator: { _id: creator._id, name: creator.name }
       }); 
     })
     .catch(err => {
@@ -141,6 +154,12 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error; 
       }
+      // check if current ID matches the created user ID
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized');
+        error.statusCode = 403;
+        throw error;   
+      }
       // Delete old image if url path is not the same
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl); 
@@ -173,16 +192,33 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      // check if current ID matches the created user ID
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized');
+        error.statusCode = 403;
+        throw error;
+      }
       // check if the creator is the currently logged in user
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId); 
     })
     .then(result => {
-      console.log(result);
-      res.status(200).json({ message: 'POST DELETED SUCCESSFULLY!' });
+      // Find current user 
+      return User.findByIdAndRemove(req.userId);
+    })
+    .then(user => {
+      // use mongoose method to remove post reference
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(result => {
+      res.status(200).json({ message: "POST DELETED SUCCESSFULLY!" });
     })
     .catch(err => {
-
+      if (!err.statusCode) {
+        err.statusCode = 500;
+      }
+      next(err); 
     });
 }
